Create two new modules for mockmvc and restassured support
Remove the dependency to a specific spring-restdocs client from
the restdocs-api-spec module and create two new modules:
- restdocs-api-spec-mockmvc
  Contains the wrapper for spring-restdocs-mockmvnc
- restdocs-api-spec-restassured
  Contains the wrapper for spring-restdocs-restassured
