+++ settings.gradle
@@ -7,3 +7,5 @@ include 'restdocs-api-spec-openapi3-generator'
 include 'restdocs-api-spec-gradle-plugin'
 include 'restdocs-api-spec-sample'
 project(':restdocs-api-spec-sample').projectDir = file('samples/restdocs-api-spec-sample')
+include 'restdocs-api-spec-mockmvc'
+include 'restdocs-api-spec-restassured'
